requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
python-dotenv==0.19.0
requests==2.26.0
cryptography==3.4.7
flask-cors==5.0.0
py-clob-client==0.18.0
kalshi-python==2.0.0
scikit-learn==1.5.2
scipy==1.7.1
supabase==1.0.3